Make Tradeshift the owner of this fork
This only updates the Maven information. The most notable changes is,
that we want to use our own `groupId`.

We also remove Blue Apron from CODEOWNERS, as it is not applicable
inside the Tradeshift GitHub organisation.

The log4j.properties is not referenced anywhere, and the settings.xml
depends on Blue Apron build environment.